refactor: move local decoding to core::decoding
