+++ src/core/decoding/modules/mod.rs
@@ -1,5 +1,6 @@
 use crate::{core::decoding::reader::WasmDecoder, DecodingError};
 
+pub mod code_section;
 pub mod custom_section;
 pub mod indices;
 pub mod sections;
